feat(ble): add whitelist address support for device discovery
Add optional whitelist_address parameter to parse_scan_response to allow
discovery of devices that don't advertise the service UUID. When a target
address is specified, the scan will include all devices and filter based on
the whitelist match.

This enables discovery of devices that may not be advertising the expected
service UUID but are known by their address or name.